feat/keyp-350 rename examples
