+++ java/component/keyple-plugin/remote/example/common/build.gradle
@@ -1,7 +1,7 @@
 apply plugin: 'java'
 apply plugin: 'java-library'
 
-project.description = 'Example of Keyple Plugin Remote SE - common'
+project.description = 'Example of Keyple Plugin Remote - common'
 
 
 repositories {
@@ -13,8 +13,8 @@ dependencies {
 
     /* Keyple dependencies */
     implementation group: 'org.eclipse.keyple', name: 'keyple-java-calypso', version: keyple_version
-    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remotese-nativese', version: keyple_version
-    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remotese-virtualse', version: keyple_version
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remote-native', version: keyple_version
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remote-virtual', version: keyple_version
 
     /* Keyple plugin dependencies */
     implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-stub', version: keyple_version
