+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-service/build.gradle
@@ -1,0 +1,46 @@
+plugins {
+    id 'java'
+    id 'io.quarkus'
+}
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+}
+
+dependencies {
+
+    /* Keyple dependencies */
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-calypso', version: keyple_version
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remote-native', version: keyple_version
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-remote-virtual', version: keyple_version
+
+    /* Keyple plugin dependencies */
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-stub', version: keyple_version
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-pcsc', version: keyple_version
+
+    /* Example dependencies */
+    implementation project(path: ':common')
+
+    /* Quarkus */
+    implementation 'io.quarkus:quarkus-resteasy-jsonb'
+    implementation 'io.quarkus:quarkus-resteasy'
+    implementation 'io.quarkus:quarkus-rest-client'
+    implementation enforcedPlatform("${quarkusPlatformGroupId}:${quarkusPlatformArtifactId}:${quarkusPlatformVersion}")
+
+    testImplementation 'io.quarkus:quarkus-junit5'
+    testImplementation 'io.rest-assured:rest-assured'
+}
+
+group 'org.eclipse.keyple.example.calypso.remote'
+version '1.0.0-SNAPSHOT'
+
+compileJava {
+    options.encoding = 'UTF-8'
+    options.compilerArgs << '-parameters'
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
