+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-service/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Oct 19 10:48:40 CEST 2020
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.6.1-all.zip
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
