+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-service/settings.gradle
@@ -1,0 +1,14 @@
+pluginManagement {
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        gradlePluginPortal()
+    }
+    plugins {
+        id 'io.quarkus' version "${quarkusPluginVersion}"
+    }
+}
+rootProject.name = 'keyple-example-remote-web-service'
+
+include ':common'
+project(':common').projectDir = new File("../common")
