+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-service/src/main/docker/Dockerfile.native
