+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-socket/build.gradle
