+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-socket/gradle/wrapper/gradle-wrapper.properties
