+++ java/component/keyple-plugin/remote/example/keyple-example-remote-web-socket/src/main/resources/application.properties
