--- java/component/keyple-plugin/remote/example/keyple-example-remotese-web-service/settings.gradle
@@ -1,14 +1,0 @@
-pluginManagement {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        gradlePluginPortal()
-    }
-    plugins {
-        id 'io.quarkus' version "${quarkusPluginVersion}"
-    }
-}
-rootProject.name = 'keyple-example-remotese-web-service'
-
-include ':common'
-project(':common').projectDir = new File("../common")
