--- java/component/keyple-plugin/remote/example/keyple-example-remotese-web-service/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-remotese-plugin-api/mp-rest/url=http://0.0.0.0:8080/
-remotese-plugin-api/mp-rest/scope=javax.inject.Singleton
-quarkus.index-dependency.keyple.group-id=org.eclipse.keyple
-quarkus.index-dependency.keyple.artifact-id=keyple-java-plugin-remotese-core
-quarkus.package.type=uber-jar
-quarkus.log.level=INFO
-quarkus.log.category."org.eclipse.keyple".level=TRACE
